# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym Equipment Directory API",
    "description": "ジムと設備の検索API（MVP）。\n- アプリは postgresql+asyncpg（非同期）\n- Alembic は postgresql+psycopg2（同期）\n- `gyms.last_verified_at_cached` はトリガーで自動更新\n",
    "version": "0.1.0"
  },
  "paths": {
    "/gyms/search": {
      "get": {
        "tags": [
          "gyms",
          "gyms"
        ],
        "summary": "ジム検索（設備フィルタ + Keysetページング）",
        "description": "都道府県/市区町村スラッグ、設備スラッグ（CSV）、緯度経度+半径でフィルタします。\n- sort=freshness: gyms.last_verified_at_cached DESC, id ASC\n- sort=richness: GymEquipment をスコア合算し降順\n （1.0 + min(count,5)*0.1 + min(max_weight_kg/60,1.0)*0.1）\n - sort=score: freshness(0.6)とrichness(0.4)を合算した最終スコア降順\n- equipment_match=all の場合、指定スラッグを**すべて**含むジムのみ返します\n- sort=gym_name: name ASC, id ASC（Keyset）\n- sort=created_at: created_at DESC, id ASC（Keyset）\n- sort=distance: 指定座標からのHaversine距離 ASC, id ASC（lat/lng 必須）",
        "operationId": "search_gyms_gyms_search_get",
        "parameters": [
          {
            "name": "pref",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "都道府県スラッグ（lower）例: chiba",
              "title": "Pref"
            },
            "description": "都道府県スラッグ（lower）例: chiba"
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "市区町村スラッグ（lower）例: funabashi",
              "title": "City"
            },
            "description": "市区町村スラッグ（lower）例: funabashi"
          },
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 90.0,
                  "minimum": -90.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "検索基準点の緯度（度）",
              "title": "Lat"
            },
            "description": "検索基準点の緯度（度）"
          },
          {
            "name": "lng",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 180.0,
                  "minimum": -180.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "検索基準点の経度（度）",
              "title": "Lng"
            },
            "description": "検索基準点の経度（度）"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "検索半径（km）",
              "title": "Radius Km"
            },
            "description": "検索半径（km）"
          },
          {
            "name": "equipments",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "設備スラッグCSV（例: squat-rack,dumbbell）",
              "title": "Equipments"
            },
            "description": "設備スラッグCSV（例: squat-rack,dumbbell）"
          },
          {
            "name": "equipment_match",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "any"
              ],
              "type": "string",
              "description": "equipments の一致条件",
              "default": "all",
              "title": "Equipment Match"
            },
            "description": "equipments の一致条件"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "freshness",
                "richness",
                "gym_name",
                "created_at",
                "score",
                "distance"
              ],
              "type": "string",
              "description": "並び順",
              "default": "score",
              "title": "Sort"
            },
            "description": "並び順"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ページ番号（1始まり）",
              "default": 1,
              "title": "Page"
            },
            "description": "ページ番号（1始まり）"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "1ページ件数（1..100）",
              "examples": [
                10
              ],
              "title": "Page Size"
            },
            "description": "1ページ件数（1..100）"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "1ページ件数（互換用, 1..100）",
              "title": "Per Page"
            },
            "description": "1ページ件数（互換用, 1..100）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "limit（互換用, 1..100）",
              "title": "Limit"
            },
            "description": "limit（互換用, 1..100）"
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyset 継続トークン（互換用）",
              "title": "Page Token"
            },
            "description": "Keyset 継続トークン（互換用）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymSearchPageDTO"
                }
              }
            }
          },
          "422": {
            "description": "validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/nearby": {
      "get": {
        "tags": [
          "gyms",
          "gyms"
        ],
        "summary": "ジム近傍検索（Haversine + Keyset）",
        "description": "指定した座標から半径 `radius_km` 以内のジムを、\nHaversine距離の昇順・id昇順で返します。ページングは距離+idのKeysetです。",
        "operationId": "gyms_nearby_gyms_nearby_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "緯度",
              "title": "Lat"
            },
            "description": "緯度"
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "description": "経度",
              "title": "Lng"
            },
            "description": "経度"
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0.0,
              "description": "検索半径（km）",
              "default": 5.0,
              "title": "Radius Km"
            },
            "description": "検索半径（km）"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "ページ番号（1始まり）",
              "default": 1,
              "title": "Page"
            },
            "description": "ページ番号（1始まり）"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "1ページ件数（1..100）",
              "title": "Page Size"
            },
            "description": "1ページ件数（1..100）"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "1ページ件数（互換用, 1..100）",
              "title": "Per Page"
            },
            "description": "1ページ件数（互換用, 1..100）"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "limit（互換用, 1..100）",
              "title": "Limit"
            },
            "description": "limit（互換用, 1..100）"
          },
          {
            "name": "page_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Keyset継続トークン（互換用）",
              "title": "Page Token"
            },
            "description": "Keyset継続トークン（互換用）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymNearbyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{slug}": {
      "get": {
        "tags": [
          "gyms",
          "gyms"
        ],
        "summary": "ジム詳細を取得",
        "description": "ジム詳細を返却します。`include=score` を指定すると freshness/richness/score を同梱します。",
        "operationId": "get_gym_detail_gyms__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "例: include=score",
              "title": "Include"
            },
            "description": "例: include=score"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "ジムが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/by-id/{canonical_id}": {
      "get": {
        "tags": [
          "gyms",
          "gyms"
        ],
        "summary": "ジム詳細を canonical_id で取得",
        "description": "canonical UUID を用いてジムの詳細を返却します。`include=score` を指定すると freshness/richness/score を同梱します。",
        "operationId": "get_gym_detail_by_id_gyms_by_id__canonical_id__get",
        "parameters": [
          {
            "name": "canonical_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Canonical Id"
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "例: include=score",
              "title": "Include"
            },
            "description": "例: include=score"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GymDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "ジムが見つかりません",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gyms/{slug}/report": {
      "post": {
        "tags": [
          "gyms",
          "gyms"
        ],
        "summary": "誤り報告を送信",
        "operationId": "report_gym_gyms__slug__report_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "エラーメッセージ"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "examples": [
          {
            "detail": "Not Found"
          }
        ]
      },
...
      "GymDetailDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ジムID"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "ジムスラッグ"
          },
          "canonical_id": {
            "type": "string",
            "title": "Canonical Id",
            "description": "ジムの canonical UUID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "名称"
          },
          "city": {
            "type": "string",
            "title": "City",
            "description": "市区町村スラッグ"
          },
          "pref": {
            "type": "string",
            "title": "Pref",
            "description": "都道府県スラッグ"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address",
            "description": "住所（任意）"
          },
          "official_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Official Url",
            "description": "公式サイトURL（任意）"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude",
            "description": "緯度（任意）"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude",
            "description": "経度（任意）"
          },
          "last_verified_at_cached": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At Cached",
            "description": "最終確認日時（UTC, nullable）"
          },
          "equipments": {
            "items": {
              "$ref": "#/components/schemas/GymEquipmentLineDTO"
            },
            "type": "array",
            "title": "Equipments",
            "description": "設備一覧（JOIN済み）"
          },
          "gym_equipments": {
            "items": {
              "$ref": "#/components/schemas/GymEquipmentSummaryDTO"
            },
            "type": "array",
            "title": "Gym Equipments",
            "description": "設備ごとの在/無・検証状況などの詳細サマリ"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/GymImageDTO"
            },
            "type": "array",
            "title": "Images",
            "description": "関連画像の一覧"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "設備情報の最終更新（= last_verified_at の最大）"
          },
          "freshness": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness",
            "description": "0..1"
          },
          "richness": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Richness",
            "description": "0..1"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "0..1"
          },
          "equipment_details": {
            "items": {
              "$ref": "#/components/schemas/GymEquipmentLineDTO"
            },
            "type": "array",
            "title": "Equipment Details",
            "description": "旧API互換のため equipment_details を提供する。",
            "readOnly": true
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At",
            "description": "旧フィールド名の互換用。",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "canonical_id",
          "name",
          "city",
          "pref",
          "equipments",
          "equipment_details",
          "last_verified_at"
        ],
        "title": "GymDetailDTO",
        "description": "Full detail payload for a gym.",
        "examples": [
          {
            "address": "千葉県船橋市本町1-2-3",
            "canonical_id": "11111111-2222-3333-4444-555555555555",
            "city": "funabashi",
            "equipments": [
              {
                "category": "strength",
                "count": 2,
                "equipment_name": "スクワットラック",
                "equipment_slug": "squat-rack",
                "max_weight_kg": 180
              }
            ],
            "freshness": 0.9,
            "gym_equipments": [
              {
                "availability": "present",
                "category": "strength",
                "count": 2,
                "last_verified_at": "2025-09-01T12:34:56Z",
                "max_weight_kg": 180,
                "name": "スクワットラック",
                "slug": "squat-rack",
                "source": "https://example.com/source",
                "verification_status": "verified"
              }
            ],
            "id": 1,
            "images": [
              {
                "created_at": "2025-09-01T12:34:56Z",
                "source": "instagram",
                "url": "https://example.com/image.jpg",
                "verified": false
              }
            ],
            "last_verified_at_cached": "2025-09-01T12:34:56Z",
            "latitude": 35.7001,
            "longitude": 139.9823,
            "name": "Awesome Gym",
            "official_url": "https://awesome-gym.example.com",
            "pref": "chiba",
            "richness": 0.7,
            "score": 0.82,
            "slug": "awesome-gym",
            "updated_at": "2025-09-01T12:34:56Z"
          }
        ]
      },
...
      "GymNearbyResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GymNearbyItem"
            },
            "type": "array",
            "title": "Items",
            "description": "検索結果"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "総件数",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "現在のページ（1始まり）",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "1ページ件数",
            "default": 20
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "次ページが存在するか",
            "default": false
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev",
            "description": "前ページが存在するか",
            "default": false
          },
          "page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Token",
            "description": "継続トークン（Keyset互換用、未使用時はnull）"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "GymNearbyResponse"
      },
      "GymSearchPageDTO": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GymSummaryDTO"
            },
            "type": "array",
            "title": "Items",
            "description": "検索結果"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "総件数",
            "default": 0
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "現在のページ（1始まり）",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "1ページ件数",
            "default": 20
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "次ページが存在するか",
            "default": false
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev",
            "description": "前ページが存在するか",
            "default": false
          },
          "page_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Token",
            "description": "継続トークン（Keyset互換用、未使用時はnull）"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "GymSearchPageDTO",
        "description": "Search result page for gyms including pagination metadata.",
        "examples": [
          {
            "has_more": true,
            "has_prev": false,
            "items": [
              {
                "canonical_id": "11111111-2222-3333-4444-555555555555",
                "city": "funabashi",
                "freshness_score": 0.93,
                "id": 1,
                "last_verified_at": "2025-09-01T12:34:56Z",
                "name": "Dummy Gym",
                "official_url": "https://dummy-gym.example.com",
                "pref": "chiba",
                "richness_score": 0.68,
                "score": 0.84,
                "slug": "dummy-gym"
              }
            ],
            "page": 1,
            "page_size": 20,
            "total": 10
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ]
}